refactor(surveys): colocate useNoOverlayModal with its only caller [ENG-2334]
62bd3e1 put the hook in lib/use-no-overlay-modal.ts, which fixed S3776 but
failed the Sonar quality gate: 47.7% coverage on new code against a required
80%. `**/*.tsx` is coverage-excluded, `.ts` is not, so a 22-line hook with no
unit tests became 22 uncovered new lines. Traded one Sonar issue for a worse
one.

Moved to module scope in survey-container.tsx, which keeps the S3776 fix
intact — cognitive complexity is measured per function, so what reduces
SurveyContainer's score is the code leaving the component, not the file it
lands in.

It belongs here on its own merits, not only for the metric: it has exactly
one call site and is specific to this component's modal semantics, so it is
not shared code, and being private now says so. Its whole behaviour is DOM
listeners and a timer, which per AGENTS.md is covered by Playwright rather
than unit tests — as a lib module it would have needed 22 lines of the kind
of test the repo declines to write. The reasoning is recorded at the hook.

Behaviour re-verified against a fresh bundle after the move, both modal
paths: no overlay announces and closes on Escape, dark overlay stays silent
and closes via the focus trap, both keep the dialog's accessible name and
description and the single h1.
